backend/requirements.txt
# ===== CORE DEPENDENCIES =====
# Web Framework & API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# File Uploads (required for FastAPI file handling)
python-multipart>=0.0.6

# Data Processing
numpy>=1.24.0

# ===== DATA PROCESSING =====
# Image processing (basic operations only)
Pillow>=10.0.0

# ===== GPU SUPPORT (Optional - uncomment if you have CUDA GPU) =====
# For GPU acceleration, uncomment these and comment out CPU versions above:
# torch>=2.0.0  # CUDA version (slower builds)
# torchvision>=0.15.0  # CUDA version (slower builds)

# ===== DEVELOPMENT DEPENDENCIES =====
# Testing
pytest>=7.4.0

# Code formatting & linting
black>=23.0.0
flake8>=6.0.0

# ===== NOTES =====
# - All dependencies are pinned to minimum versions for compatibility
# - CPU versions work on all systems, GPU versions require CUDA
# - First run may download AI models (~500MB), ensure stable internet
# - System requires ~2GB free disk space for AI models
